`Refactor chat feature: remove unused components and files, update chat header and list, and modify conversational panel and message input`
