--- src/app/features/chat/chat-header/chat-header.html
@@ -1,56 +1,0 @@
-<div
-  [class]="`h-16 flex items-center justify-between px-4 border-b ${darkMode ? 'bg-gray-800 border-gray-700' : 'bg-white border-gray-200'}`"
->
-  <div class="flex items-center">
-    <div class="relative">
-      <img
-        class="h-10 w-10 rounded-full"
-        src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-        alt="Your avatar"
-      />
-      <span
-        class="absolute bottom-0 right-0 block h-2.5 w-2.5 rounded-full bg-green-400 ring-2 ring-white"
-      ></span>
-    </div>
-    <div class="ml-3">
-      <p [class]="`font-medium ${darkMode ? 'text-white' : 'text-gray-900'}`">
-        Your Name
-      </p>
-      <p [class]="`text-xs ${darkMode ? 'text-gray-400' : 'text-gray-500'}`">
-        Online
-      </p>
-    </div>
-  </div>
-  <div class="flex items-center space-x-3">
-    <button
-      class="`p-2 rounded-full ${darkMode ? 'hover:bg-gray-700' : 'hover:bg-gray-100'}`"
-    >
-      <i
-        [class]="`pi pi-search h-5 w-5 ${darkMode ? 'text-gray-400' : 'text-gray-500'}`"
-      ></i>
-      <!-- <SearchIcon
-            [class]=""
-          /> -->
-    </button>
-    <button
-      [class]="
-        'p-2 rounded-sm ' +
-        (darkMode ? 'bg-gray-800 text-gray-300' : 'bg-white text-gray-600')
-      "
-      (click)="toggleDarkMode()"
-    >
-      @if (darkMode) {
-      <i class="pi pi-sun h-5 w-5"></i>
-      }@else {
-      <i class="pi pi-moon h-5 w-5"></i>
-      }
-    </button>
-    <button
-      [class]="`p-2 rounded-full ${darkMode ? 'hover:bg-gray-700' : 'hover:bg-gray-100'}`"
-    >
-      <i
-        [class]="`pi pi-ellipsis-v h-5 w-5 ${darkMode ? 'text-gray-400' : 'text-gray-500'}`"
-      ></i>
-    </button>
-  </div>
-</div>
