--- src/app/features/chat/contact-list/contact-list.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ContactList } from './contact-list';
-
-describe('ContactList', () => {
-  let component: ContactList;
-  let fixture: ComponentFixture<ContactList>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ContactList]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(ContactList);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
