--- src/app/features/chat/conversational-panel/chat-message/chat-message.html
@@ -1,45 +1,0 @@
-<div [class]="`flex ${message.isUser ? 'justify-end' : 'justify-start'}`">
-  @if (!message.isUser) {
-
-  <img
-    class="h-8 w-8 rounded-full mr-2 self-end"
-    [src]="message.avatar"
-    [alt]="message.sender"
-  />
-  }
-  <div class="flex flex-col">
-    <div
-      [class]="`p-3 rounded-2xl max-w-md ${message.isUser ? (darkMode ? 'bg-indigo-600 text-white' : 'bg-indigo-500 text-white') : darkMode ? 'bg-gray-800 text-white' : 'bg-white text-gray-800'}`"
-    >
-      <p>{{ message.message }}</p>
-    </div>
-    <div
-      [class]="`flex items-center mt-1 ${message.isUser ? 'justify-end' : 'justify-start'}`"
-    >
-      <span [class]="`text-xs ${darkMode ? 'text-gray-400' : 'text-gray-500'}`">
-        {{ message.time }}
-      </span>
-      @if(message.isUser){
-
-      <div class="ml-1">
-        @if(message.read){
-        <div class="flex -space-x-1">
-          <i class="pi pi-check h-3 w-3 text-blue-500"></i>
-          <i class="pi pi-check h-3 w-3 text-blue-500"></i>
-        </div>
-        }@else{
-        <i class="pi pi-check h-3 w-3 text-gray-400"></i>
-        }
-      </div>
-
-      }
-    </div>
-  </div>
-  @if(message.isUser){
-  <img
-    class="h-8 w-8 rounded-full ml-2 self-end"
-    [src]="message.avatar"
-    [alt]="message.sender"
-  />
-  }
-</div>
