--- src/app/features/chat/conversational-panel/chat-message/chat-message.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ChatMessage } from './chat-message';
-
-describe('ChatMessage', () => {
-  let component: ChatMessage;
-  let fixture: ComponentFixture<ChatMessage>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ChatMessage]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(ChatMessage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
