--- src/app/features/chat/conversational-panel/conversational-panel.html
@@ -1,25 +1,0 @@
-<div [class]="`flex-grow p-4 overflow-y-auto ${darkMode ? 'bg-gray-900' : 'bg-gray-50'}`">
-  <div class="max-w-3xl mx-auto">
-    <div class="space-y-4">
-      @for (message of currentChat; track $index) {
-      <app-chat-message [message]="message" />
-      } @if (typing && activeChat === 1) {
-      <div class="flex items-center">
-        <img class="h-8 w-8 rounded-full mr-2"
-          src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-          alt="Sarah Johnson" />
-        <div [class]="`p-3 rounded-2xl max-w-md ${darkMode ? 'bg-gray-800' : 'bg-white'}`">
-          <div class="flex space-x-1">
-            <div [class]="`w-2 h-2 rounded-full animate-bounce ${darkMode ? 'bg-gray-500' : 'bg-gray-400'}`"
-              style="animation-delay: 0ms"></div>
-            <div [class]="`w-2 h-2 rounded-full animate-bounce ${darkMode ? 'bg-gray-500' : 'bg-gray-400'}`"
-              style="animation-delay: 150ms"></div>
-            <div [class]="`w-2 h-2 rounded-full animate-bounce ${darkMode ? 'bg-gray-500' : 'bg-gray-400'}`"
-              style="animation-delay: 300ms"></div>
-          </div>
-        </div>
-      </div>
-      }
-    </div>
-  </div>
-</div>
