--- src/app/features/chat/message-input/message-input.html
@@ -1,56 +1,0 @@
-<div
-  [class]="`p-4 border-t ${darkMode ? 'bg-gray-800 border-gray-700' : 'bg-white border-gray-200'}`"
->
-  @if (showEmojiPicker) {
-  <div
-    [class]="`grid grid-cols-8 gap-2 p-3 mb-2 rounded-lg ${darkMode ? 'bg-gray-700' : 'bg-gray-100'}`"
-  >
-    @for (emoji of emojis;track $index; let index = $index) {
-    <button
-      key="{index}"
-      class="text-xl hover:bg-opacity-50 hover:bg-gray-200 rounded p-1"
-      (click)="addEmoji(emoji)"
-    >
-      {{ emoji }}
-    </button>
-    }
-  </div>
-  }
-  <form (ngSubmit)="handleSubmit()" class="flex items-center">
-    <button
-      type="button"
-      [class]="`p-2 rounded-sm ${darkMode ? 'hover:bg-gray-700 text-gray-400' : 'hover:bg-gray-100 text-gray-500'}`"
-      (click)="toggleEmojiPicker()"
-    >
-      <i class="pi pi-face-smile h-5 w-5"></i>
-    </button>
-    <button
-      type="button"
-      [class]="`p-2 rounded-sm ${darkMode ? 'hover:bg-gray-700 text-gray-400' : 'hover:bg-gray-100 text-gray-500'}`"
-    >
-      <i class="pi pi-paperclip h-5 w-5"></i>
-    </button>
-    <input
-      [formControl]="message"
-      (focus)="onFocus.emit(true)"
-      (blur)="onFocus.emit(false)"
-      placeholder="Type a message..."
-      onFocus="{onFocus}"
-      [class]="`flex-grow mx-2 p-2 rounded-sm focus:outline-none focus:ring-2 focus:ring-indigo-500 ${darkMode ? 'bg-gray-700 text-white placeholder-gray-400 border-gray-600' : 'bg-gray-100 text-gray-900 placeholder-gray-500 border-gray-200'}`"
-      placeholder="Type a message..."
-    />
-    <button
-      type="button"
-      [class]="`p-2 rounded-sm ${darkMode ? 'hover:bg-gray-700 text-gray-400' : 'hover:bg-gray-100 text-gray-500'}`"
-    >
-      <i class="pi pi-microphone h-5 w-5"></i>
-    </button>
-    <button
-      type="submit"
-      [disabled]="!message.value?.trim()"
-      [class]="`p-2 rounded-sm ${message.value?.trim() ? 'bg-indigo-600 hover:bg-indigo-700 text-white' : darkMode ? 'text-gray-500' : 'text-gray-400'}`"
-    >
-      <i class="pi pi-send h-5 w-5"></i>
-    </button>
-  </form>
-</div>
